requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
idna==2.10
multidict==5.1.0
pycodestyle==2.7.0
python-engineio==3.14.2
python-socketio==4.6.1
requests==2.25.1
six==1.15.0
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.26.4
websocket-client==0.58.0
websockets==8.1
yarl==1.6.3
